prj_summer-badge/tool-py_db-tools/requirements.txt
mysql-connector-python==8.0.33
pandas==2.0.2
